Sockets stream and dgram fully functional
